Refactor resource handling to standardize fileUrl usage
- Updated API routes and components to replace 'url' with 'fileUrl' for consistency in resource management.
- Enhanced parameter naming in API documentation for clarity.
- Removed unused code and logging related to resource state management in the module detail page.
- Streamlined resource filtering and state management by utilizing a custom hook for better performance and maintainability.
